updated URL and MySQL Injection Protection

+++ admin/includes/view_all_users.php
@@ -87,11 +87,18 @@
 
 
     if(isset($_GET['delete'])){
-      $the_user_id = $_GET['delete'];
-      
-      $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-      $delete_query = mysqli_query($connection, $query);
-      header("Location: users.php"); //header
+      if(isset($_SESSION['user_role'])){
+        if($_SESSION['user_role'] == 'admin'){
+         
+          $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+          
+          $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+          $delete_query = mysqli_query($connection, $query);
+          header("Location: users.php"); //header
+
+        }
+      }
+
     }
   
   ?>
